move state to separate file
